refactor: 플레이어가 소유한 카드 네임과 카드 목록 변수명 변경
