+++ src/main/java/blackjack/domain/player/Dealer.java
@@ -1,6 +1,6 @@
 package blackjack.domain.player;
 
-import blackjack.domain.card.Deck;
+import blackjack.domain.card.PlayingCards;
 import blackjack.domain.result.Match;
 
 public class Dealer extends AbstractPlayer implements Player {
@@ -10,7 +10,7 @@ public class Dealer extends AbstractPlayer implements Player {
     public static final int EXCEED_POINT = 17;
 
     public Dealer() {
-        this.playingCards = new Deck();
+        this.playingCards = new PlayingCards();
         this.name = NAME;
     }
